Add date modification parsing
