+++ src/uucore/src/lib/parser.rs
@@ -2,7 +2,6 @@
 //
 // For the full copyright and license information, please view the LICENSE
 // file that was distributed with this source code.
-pub mod parse_date_modifier;
 pub mod parse_glob;
 pub mod parse_size;
 pub mod parse_time;
